third_party/WebKit/LayoutTests/editing/deleting/merge-paragraph-into-pre.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable><pre>hello</pre>\n|world</div>',
        'delete',
        // We should not cancel styles added BY PRE by those of document's
        // default style.
        '<div contenteditable><pre>hello|world</pre></div>'),
        'Delete line break after PRE');
</script>
