third_party/WebKit/LayoutTests/editing/deleting/merge-paragraph-into-pre-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
This tests deleting line break after pre.
WebKit should not cancel styles added by pre by those of document's default style.
| "
"
| <pre>
|   "hello<#selection-caret>world"
